cog-services-python/vision/requirements.txt
azure-cognitiveservices-vision-computervision>=0.3.0
requests
Pillow
python-dotenv